Get correct Capabilities type from Options
Change AnnotationsReader to allow to get the annotated Options of a
given type, to allow multiple configuration Options in the same test
class. Previously it would lead to a ClassCastException since it assumed
a single annotated Options.
Update DriverHandler implementations to request the corresponding type
(e.g. EdgeOptions for EdgeDriverHandler).