Let HtmlUnitDriver be chosen as a substitute to Selenium WebDriver
In the past I needed to change the signature of my unittest to be able to run it with HtmlUnit instead of with a slow browser.
As HtmlUnitDriver implements WebDriver it seems logical to be able to use it as a pretend Selenium WebDriver when adding "-Dwdm.defaultBrowser=HTMLUNIT" to the command line.